requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
google-auth==2.26.1
